refactor: added generics to the radio list item
