+++ lib/core/common/ui/forms/flex_radio_list.dart
@@ -1,60 +1,14 @@
-import 'package:flutter_template_core/core/common/ui/components/large_button.dart';
-import 'package:flutter_template_core/core/extensions/ui_extensions.dart';
+import 'package:flutter_template_core/core/common/ui/components/flex_radio_list_item.dart';
 import 'package:flutter_template_core/core/theme/app_layout.dart';
 import 'package:fpdart/fpdart.dart';
 import 'package:reactive_forms_annotations/reactive_forms_annotations.dart';
 
-class RadioListItemClass {
-  const RadioListItemClass({required this.name});
-  final String name;
-}
-
-class RadioListItem extends StatelessWidget {
-  const RadioListItem({
-    required this.name,
-    required this.icon,
-    this.selected = false,
-    this.onPressed,
-    super.key,
-  });
-
-  final String name;
-  final IconData icon;
-
-  final bool selected;
-  final VoidCallback? onPressed;
-
-  @override
-  Widget build(BuildContext context) {
-    return LargeButton(
-      onPressed: onPressed,
-      borderColor:
-          selected ? context.colors.foregroundPrimary : context.colors.divider,
-      borderWidth: selected ? 2 : 1,
-      child: Padding(
-        padding: const EdgeInsets.symmetric(
-          horizontal: AppLayout.p4,
-          vertical: AppLayout.p6,
-        ),
-        child: Row(
-          children: [
-            Icon(icon),
-            const SizedBox(width: AppLayout.p3),
-            Text(
-              name,
-              style: context.typography.bodyMedium
-                  .copyWith(fontWeight: FontWeight.w600),
-            ),
-          ],
-        ),
-      ),
-    );
-  }
-}
-
-class FlexRadioList extends ReactiveFormField<int, int> {
+class FlexRadioList<T extends RadioListItemMixin>
+    extends ReactiveFormField<int, int> {
   FlexRadioList({
-    required List<RadioListItem> options,
+    required List<T> options,
+    // ignore: avoid_positional_boolean_parameters
+    required T Function(T, bool, VoidCallback) builder,
     super.key,
     super.formControlName,
     super.formControl,
@@ -72,14 +26,9 @@ class FlexRadioList extends ReactiveFormField<int, int> {
                       .mapWithIndex(
                         (e, index) => Column(
                           children: [
-                            RadioListItem(
-                              name: e.name,
-                              icon: e.icon,
-                              selected: field.value == index,
-                              onPressed: () {
-                                field.didChange(index);
-                              },
-                            ),
+                            builder(e, field.value == index, () {
+                              field.didChange(index);
+                            }),
                             const SizedBox(height: AppLayout.p2),
                           ],
                         ),
